Split ADO int tests up into 2 classes (one per test)
